fix(cli): split doctor dev-source helpers out; regenerate CLI reference
Keep src/teatree/cli/doctor/app.py under the module-health LOC cap after
the run_doctor_checks exit-code split by moving the editable dev-source
override helpers (_find_host_project_root, _find_teatree_pyproject_from_cwd,
_patch_uv_source, _write_dev_sources_marker, _DEV_HIDDEN_FILES) into a new
dev_sources module and re-exporting them; drop the unused _DEV_SOURCES_FILE.

Regenerate docs/generated/cli-reference.md so the docs-drift gate is green
after the merged validate-mr/slack-check help-text updates.
